+++ src/component/Dropdown/index.js
@@ -1,4 +1,4 @@
-import Dropdown from './DropDown';
+import Dropdown from './Dropdown';
 import './_dropdown.scss';
 
 export default Dropdown;